Errors feature

+++ FlexSearch.Api/Controllers/IndexController.cs
@@ -8,7 +8,8 @@ using Core.Storage;
 using Core.Storage.Database;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using SearchApi.Mappings;
+using SearchApi.Dtos;
+using SearchApi.Dtos.Mappings;
 using SearchApi.Services;
 
 
@@ -30,7 +31,7 @@ namespace SearchApi.Controllers
         public async Task<IActionResult> GetDatabases()
         {
             if (await UserService.CheckAuthorize(Request) is null)
-                return Unauthorized();
+                return Unauthorized(ErrorDto.GetAuthError());
             return Ok(await DatabaseService.GetDatabases());
         }
         
@@ -50,18 +51,23 @@ namespace SearchApi.Controllers
         public async Task<IActionResult> GetDocument(string dbname, string index, string id)
         {
             if (await UserService.CheckAuthorize(Request, false, dbname) is null)
-                return Unauthorized();
-            var result = await DatabaseService.FindById(new IndexModel(dbname, index), id);
-            if (result is null)
-                return NoContent();
-            return Ok(DocumentMapper.MapToDto(result));
+                return Unauthorized(ErrorDto.GetAuthError());
+            if (Guid.TryParse(id, out var guid))
+            {
+                var result = await DatabaseService.FindById(new IndexModel(dbname, index), guid);
+                if (result is null)
+                    return NoContent();
+                return Ok(DocumentMapper.MapToDto(result));
+            }
+
+            return BadRequest(new ErrorDto(ErrorsType.SyntaxError, $"{id} is not GUID"));
         }
 
         [HttpPost("index/{dbname}/{index}/")]
         public async Task<IActionResult> CreateIndex([FromBody] object obj, string dbname, string index)
         {
             if (await UserService.CheckAuthorize(Request, false, dbname) is null)
-                return Unauthorized();
+                return Unauthorized(ErrorDto.GetAuthError());
             try
             {
                 var id = await DatabaseService.Insert(new IndexModel(dbname, index), obj);
@@ -69,8 +75,8 @@ namespace SearchApi.Controllers
             }
             catch (Exception ex)
             {
-                _logger.Log(LogLevel.Error, $"ERROR: Create object {obj} in {dbname}/{index}, Error: {ex.Message}");
-                return StatusCode(500);
+                _logger.Log(LogLevel.Error, $"ERROR: Create object {obj} in {dbname}/{index}, Error: {ex}");
+                return StatusCode(500, new ErrorDto(ErrorsType.SystemError, ex.Message));
             }
         }
 
@@ -78,17 +84,22 @@ namespace SearchApi.Controllers
         public async Task<IActionResult> RenameIndex(string dbname, string index, string name)
         {
             if (await UserService.CheckAuthorize(Request, false, dbname) is null)
-                return Unauthorized();
+                return Unauthorized(ErrorDto.GetAuthError());
             try
             {
                 await DatabaseService.RenameIndex(new IndexModel(dbname, index), name);
                 _logger.Log(LogLevel.Information, $"INFO: Rename index {dbname}/{index}. New name: {name}");
                 return StatusCode(202);
             }
-            catch (DirectoryNotFoundException)
+            catch (FileNotFoundException)
             {
                 _logger.Log(LogLevel.Error, $"ERROR: index {index} not found");
-                return BadRequest($"Не найдено индекса с именем {index}");
+                return BadRequest(new ErrorDto(ErrorsType.SyntaxError, $"Index {index} not found"));
+            }
+            catch (IOException)
+            {
+                _logger.Log(LogLevel.Error, $"ERROR: index {index} not found");
+                return BadRequest(new ErrorDto(ErrorsType.SyntaxError, $"Index {index} already exists"));
             }
         }
         
@@ -96,16 +107,16 @@ namespace SearchApi.Controllers
         public async Task<IActionResult> UpdateObject([FromBody] object obj,string dbname, string index, string id)
         {
             if (await UserService.CheckAuthorize(Request, false, dbname) is null)
-                return Unauthorized();
+                return Unauthorized(ErrorDto.GetAuthError());
             try
             {
                 await DatabaseService.Update(new IndexModel(dbname, index), obj, id);
                 _logger.Log(LogLevel.Information, $"INFO: Update object with id: {id} in {dbname}/{index}. New object: {obj}");
             }
-            catch (FileNotFoundException)
+            catch (Exception ex)
             {
-                _logger.Log(LogLevel.Error, $"ERROR: Object with id: {id} not found");
-                return BadRequest($"Не существует записи с id: {id}");
+                _logger.Log(LogLevel.Error, $"ERROR: error updating with id: {id}, exception: {ex}");
+                return BadRequest(new ErrorDto(ErrorsType.SystemError, $"{ex.Message}"));
             }
             return StatusCode(202);
 
@@ -115,7 +126,7 @@ namespace SearchApi.Controllers
         public async Task<IActionResult> DeleteDatabase(string dbname)
         {
             if (await UserService.CheckAuthorize(Request, true) is null)
-                return Unauthorized();
+                return Unauthorized(ErrorDto.GetAuthError());
             try
             {
                 await DatabaseService.DeleteDatabase(dbname);
@@ -134,13 +145,13 @@ namespace SearchApi.Controllers
         public async Task<IActionResult> DeleteIndex(string dbname, string index)
         {
             if (await UserService.CheckAuthorize(Request, true) is null)
-                return Unauthorized();
+                return Unauthorized(ErrorDto.GetAuthError());
             try
             {
                 await DatabaseService.DeleteIndex(new IndexModel(dbname, index));
                 _logger.Log(LogLevel.Information, $"INFO: Delete index {dbname}/{index}");
             }
-            catch (DirectoryNotFoundException)
+            catch (FileNotFoundException)
             {
                 _logger.Log(LogLevel.Error, $"ERROR: index {dbname}/{index} not found");
                 return NoContent();
@@ -154,21 +165,24 @@ namespace SearchApi.Controllers
         {
             if (await UserService.CheckAuthorize(Request, true) is null)
                 return Unauthorized();
-            try
+            if (Guid.TryParse(id, out var guid))
             {
-                _logger.Log(LogLevel.Information, $"INFO: Delete object from {dbname}/{index} with id: {id}");
-                //todo: refactor this
-                var indexModel = new IndexModel(dbname, index);
-                var model = await DatabaseService.FindById(indexModel, id);
-                await DatabaseService.Delete(indexModel, model);
-            }
-            catch (FileNotFoundException)
-            {
-                _logger.Log(LogLevel.Error, $"ERROR: Object with id: {id} not found");
-                return BadRequest($"Не существует записи с id: {id}");
+                try
+                {
+                    _logger.Log(LogLevel.Information, $"INFO: Delete object from {dbname}/{index} with id: {id}");
+                    var indexModel = new IndexModel(dbname, index);
+                    var model = await DatabaseService.FindById(indexModel, guid);
+                    await DatabaseService.Delete(indexModel, model);
+                    return Ok();
+                }
+                catch (Exception ex)
+                {
+                    _logger.Log(LogLevel.Error, $"ERROR: deleting document with id: {id}. Exception: {ex}");
+                    return BadRequest(new ErrorDto(ErrorsType.SystemError, ex.Message));
+                }
             }
 
-            return Ok();
+            return BadRequest(new ErrorDto(ErrorsType.SyntaxError, $"{id} is not GUID"));
         }
     }
 }
